Fixed issue where user info wasn't getting updated until the user logs back in.
This was caused be us storing the whole user object as the session attribute. Since this only gets set once upon login, user info won't change while user is logged in. To solve this, we store just the user ID in httpSession, then use that ID to fetch the user profile directly from the database whenever we need it.